fix: canvas 렌더링 분리 및 파일 이름 변경
