+++ static/canvas/js/render-controller-panel.js
