+++ templates/canvas.html
@@ -9,7 +9,7 @@
       // Variable from FastAPI, Ignore the error occurred below line.
       const data = {{ diagram | tojson }};
     </script>
-    <script type="module" src="/static/canvas/js/canvas.js"></script>
+    <script type="module" src="/static/canvas/js/index.js"></script>
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
   </head>
   <body>
